refactor: controllers
routlar ın daha anlaşılır ve takip edilebilir olması adına birbirinden ayrılması.
