+++ package.json
@@ -21,6 +21,7 @@
   },
   "homepage": "https://github.com/acikkaynak/deprem-io-backend#readme",
   "dependencies": {
+    "@fastify/autoload": "^5.7.1",
     "@fastify/cors": "^8.2.0",
     "@fastify/swagger": "^8.3.1",
     "@fastify/swagger-ui": "^1.3.0",
@@ -30,9 +31,9 @@
     "node-cache": "^5.1.2"
   },
   "devDependencies": {
+    "@vitest/coverage-c8": "^0.28.4",
     "nodemon": "^2.0.20",
     "rome": "^11.0.0",
-    "vitest": "^0.28.4",
-    "@vitest/coverage-c8": "^0.28.4"
+    "vitest": "^0.28.4"
   }
 }
